refactor: Tweaking the types to reduce maintenance burden
BREAKING CHANGE to type import paths. Import everything from @algorandfoundation/algokit-subscriber/types now
