+++ docs/code/README.md
@@ -7,6 +7,8 @@
 ### Modules
 
 - [index](modules/index.md)
+- [types](modules/types.md)
+- [types/arc-28](modules/types_arc_28.md)
 - [types/async-event-emitter](modules/types_async_event_emitter.md)
 - [types/block](modules/types_block.md)
 - [types/subscription](modules/types_subscription.md)
