+++ docs/code/interfaces/types_arc_28.Arc28Event.md
@@ -1,0 +1,51 @@
+[@algorandfoundation/algokit-subscriber](../README.md) / [types/arc-28](../modules/types_arc_28.md) / Arc28Event
+
+# Interface: Arc28Event
+
+[types/arc-28](../modules/types_arc_28.md).Arc28Event
+
+The definition of metadata for an ARC-28 event per https://github.com/algorandfoundation/ARCs/blob/main/ARCs/arc-0028.md#event.
+
+## Table of contents
+
+### Properties
+
+- [args](types_arc_28.Arc28Event.md#args)
+- [desc](types_arc_28.Arc28Event.md#desc)
+- [name](types_arc_28.Arc28Event.md#name)
+
+## Properties
+
+### args
+
+• **args**: \{ `desc?`: `string` ; `name?`: `string` ; `type`: `string`  }[]
+
+The arguments of the event, in order
+
+#### Defined in
+
+[types/arc-28.ts:15](https://github.com/algorandfoundation/algokit-subscriber-ts/blob/main/src/types/arc-28.ts#L15)
+
+___
+
+### desc
+
+• `Optional` **desc**: `string`
+
+Optional, user-friendly description for the event
+
+#### Defined in
+
+[types/arc-28.ts:13](https://github.com/algorandfoundation/algokit-subscriber-ts/blob/main/src/types/arc-28.ts#L13)
+
+___
+
+### name
+
+• **name**: `string`
+
+The name of the event
+
+#### Defined in
+
+[types/arc-28.ts:11](https://github.com/algorandfoundation/algokit-subscriber-ts/blob/main/src/types/arc-28.ts#L11)
