+++ docs/code/modules/types_arc_28.md
@@ -1,0 +1,12 @@
+[@algorandfoundation/algokit-subscriber](../README.md) / types/arc-28
+
+# Module: types/arc-28
+
+## Table of contents
+
+### Interfaces
+
+- [Arc28Event](../interfaces/types_arc_28.Arc28Event.md)
+- [Arc28EventGroup](../interfaces/types_arc_28.Arc28EventGroup.md)
+- [Arc28EventToProcess](../interfaces/types_arc_28.Arc28EventToProcess.md)
+- [EmittedArc28Event](../interfaces/types_arc_28.EmittedArc28Event.md)
