+++ docs/code/modules/types_subscription.md
@@ -6,13 +6,10 @@
 
 ### Interfaces
 
-- [Arc28Event](../interfaces/types_subscription.Arc28Event.md)
-- [Arc28EventGroup](../interfaces/types_subscription.Arc28EventGroup.md)
-- [Arc28EventToProcess](../interfaces/types_subscription.Arc28EventToProcess.md)
-- [EmittedArc28Event](../interfaces/types_subscription.EmittedArc28Event.md)
+- [AlgorandSubscriberConfig](../interfaces/types_subscription.AlgorandSubscriberConfig.md)
+- [CoreTransactionSubscriptionParams](../interfaces/types_subscription.CoreTransactionSubscriptionParams.md)
 - [NamedTransactionFilter](../interfaces/types_subscription.NamedTransactionFilter.md)
 - [SubscriberConfigFilter](../interfaces/types_subscription.SubscriberConfigFilter.md)
-- [SubscriptionConfig](../interfaces/types_subscription.SubscriptionConfig.md)
 - [TransactionFilter](../interfaces/types_subscription.TransactionFilter.md)
 - [TransactionSubscriptionParams](../interfaces/types_subscription.TransactionSubscriptionParams.md)
 - [TransactionSubscriptionResult](../interfaces/types_subscription.TransactionSubscriptionResult.md)
@@ -26,7 +23,7 @@
 
 ### SubscribedTransaction
 
-Ƭ **SubscribedTransaction**: `TransactionResult` & \{ `arc28Events?`: [`EmittedArc28Event`](../interfaces/types_subscription.EmittedArc28Event.md)[] ; `filtersMatched?`: `string`[] ; `inner-txns?`: [`SubscribedTransaction`](types_subscription.md#subscribedtransaction)[] ; `parentTransactionId?`: `string`  }
+Ƭ **SubscribedTransaction**: `TransactionResult` & \{ `arc28Events?`: [`EmittedArc28Event`](../interfaces/types_arc_28.EmittedArc28Event.md)[] ; `filtersMatched?`: `string`[] ; `inner-txns?`: [`SubscribedTransaction`](types_subscription.md#subscribedtransaction)[] ; `parentTransactionId?`: `string`  }
 
 The common model used to expose a transaction that is returned from a subscription.
 
@@ -37,7 +34,7 @@ Substantively, based on the Indexer  [`TransactionResult` model](https://develop
 
 #### Defined in
 
-[types/subscription.ts:68](https://github.com/algorandfoundation/algokit-subscriber-ts/blob/main/src/types/subscription.ts#L68)
+[types/subscription.ts:33](https://github.com/algorandfoundation/algokit-subscriber-ts/blob/main/src/types/subscription.ts#L33)
 
 ___
 
@@ -68,4 +65,4 @@ ___
 
 #### Defined in
 
-[types/subscription.ts:255](https://github.com/algorandfoundation/algokit-subscriber-ts/blob/main/src/types/subscription.ts#L255)
+[types/subscription.ts:190](https://github.com/algorandfoundation/algokit-subscriber-ts/blob/main/src/types/subscription.ts#L190)
