+++ src/subscriptions.ts
@@ -10,11 +10,9 @@ import {
   getBlockTransactions,
   getIndexerTransactionFromAlgodTransaction,
 } from './transform'
+import type { Arc28EventGroup, Arc28EventToProcess, EmittedArc28Event } from './types/arc-28'
 import type { Block } from './types/block'
 import type {
-  Arc28EventGroup,
-  Arc28EventToProcess,
-  EmittedArc28Event,
   NamedTransactionFilter,
   SubscribedTransaction,
   TransactionFilter,
@@ -55,7 +53,8 @@ export async function getSubscribedTransactions(
   algod: Algodv2,
   indexer?: Indexer,
 ): Promise<TransactionSubscriptionResult> {
-  const { watermark, filters, maxRoundsToSync, syncBehaviour: onMaxRounds } = subscription
+  const { watermark, filters, maxRoundsToSync: _maxRoundsToSync, syncBehaviour: onMaxRounds } = subscription
+  const maxRoundsToSync = _maxRoundsToSync ?? 500
   const currentRound = (await algod.status().do())['last-round'] as number
 
   // Pre-calculate a flat list of all ARC-28 events to process
