+++ src/types/subscription.ts
@@ -1,61 +1,26 @@
 import type { ApplicationOnComplete, TransactionResult } from '@algorandfoundation/algokit-utils/types/indexer'
 import algosdk from 'algosdk'
-import ABIValue = algosdk.ABIValue
+import { Arc28EventGroup, EmittedArc28Event } from './arc-28'
 import TransactionType = algosdk.TransactionType
 
-/**
- * The definition of metadata for an ARC-28 event per https://github.com/algorandfoundation/ARCs/blob/main/ARCs/arc-0028.md#event.
- */
-export interface Arc28Event {
-  /** The name of the event */
-  name: string
-  /** Optional, user-friendly description for the event */
-  desc?: string
-  /** The arguments of the event, in order */
-  args: Array<{
-    /** The type of the argument */
-    type: string
-    /** Optional, user-friendly name for the argument */
-    name?: string
-    /** Optional, user-friendly description for the argument */
-    desc?: string
-  }>
-}
-
-/** An ARC-28 event to be processed */
-export interface Arc28EventToProcess {
-  /** The name of the ARC-28 event group the event belongs to */
-  groupName: string
-  /** The name of the ARC-28 event that was triggered */
-  eventName: string
-  /** The signature of the event e.g. `EventName(type1,type2)` */
-  eventSignature: string
-  /** The 4-byte hex prefix for the event */
-  eventPrefix: string
-  /** The ARC-28 definition of the event */
-  eventDefinition: Arc28Event
-}
-
-/** An emitted ARC-28 event extracted from an app call log. */
-export interface EmittedArc28Event extends Arc28EventToProcess {
-  /** The ordered arguments extracted from the event that was emitted */
-  args: ABIValue[]
-  /** The named arguments extracted from the event that was emitted (where the arguments had a name defined) */
-  argsByName: Record<string, ABIValue>
-}
-
-/** Specifies a group of ARC-28 event definitions along with instructions for when to attempt to process the events. */
-export interface Arc28EventGroup {
-  /** The name to designate for this group of events. */
-  groupName: string
-  /** Optional list of app IDs that this event should apply to */
-  processForAppIds?: number[]
-  /** Optional predicate to indicate if these ARC-28 events should be processed for the given transaction */
-  processTransaction?: (transaction: TransactionResult) => boolean
-  /** Whether or not to silently (with warning log) continue if an error is encountered processing the ARC-28 event data; default = false */
-  continueOnError?: boolean
-  /** The list of ARC-28 event definitions */
-  events: Arc28Event[]
+/** The result of a single subscription pull/poll. */
+export interface TransactionSubscriptionResult {
+  /** The round range that was synced from/to */
+  syncedRoundRange: [startRound: number, endRound: number]
+  /** The current detected tip of the configured Algorand blockchain. */
+  currentRound: number
+  /** The new watermark value to persist for the next call to
+   * `getSubscribedTransactions` to continue the sync.
+   * Will be equal to `syncedRoundRange[1]`. Only persist this
+   * after processing (or in the same atomic transaction as)
+   * subscribed transactions to keep it reliable. */
+  newWatermark: number
+  /** Any transactions that matched the given filter within
+   * the synced round range. This substantively uses the [indexer transaction
+   * format](https://developer.algorand.org/docs/rest-apis/indexer/#transaction)
+   * to represent the data with some additional fields.
+   */
+  subscribedTransactions: SubscribedTransaction[]
 }
 
 /** The common model used to expose a transaction that is returned from a subscription.
@@ -76,8 +41,8 @@ export type SubscribedTransaction = TransactionResult & {
   filtersMatched?: string[]
 }
 
-/** Parameters to control a single subscription pull/poll. */
-export interface TransactionSubscriptionParams {
+/** Common parameters to control a single subscription pull/poll for both `AlgorandSubscriber` and `getSubscribedTransactions`. */
+export interface CoreTransactionSubscriptionParams {
   /** The filter(s) to apply to find transactions of interest.
    * A list of filters with corresponding names.
    *
@@ -102,24 +67,15 @@ export interface TransactionSubscriptionParams {
   filters: NamedTransactionFilter[]
   /** Any ARC-28 event definitions to process from app call logs */
   arc28Events?: Arc28EventGroup[]
-  /** The current round watermark that transactions have previously been synced to.
-   *
-   * Persist this value as you process transactions processed from this method
-   * to allow for resilient and incremental syncing.
-   *
-   * Syncing will start from `watermark + 1`.
-   *
-   * Start from 0 if you want to start from the beginning of time, noting that
-   * will be slow if `onMaxRounds` is `sync-oldest`.
-   **/
-  watermark: number
   /** The maximum number of rounds to sync for each subscription pull/poll.
+   *
+   * Defaults to 500.
    *
    * This gives you control over how many rounds you wait for at a time,
    * your staleness tolerance when using `skip-sync-newest` or `fail`, and
    * your catchup speed when using `sync-oldest`.
    **/
-  maxRoundsToSync: number
+  maxRoundsToSync?: number
   /** If the current tip of the configured Algorand blockchain is more than `maxRoundsToSync`
    * past `watermark` then how should that be handled:
    *  * `skip-sync-newest`: Discard old blocks/transactions and sync the newest; useful
@@ -187,50 +143,29 @@ export interface TransactionFilter {
   arc28Events?: { groupName: string; eventName: string }[]
 }
 
-/** The result of a single subscription pull/poll. */
-export interface TransactionSubscriptionResult {
-  /** The round range that was synced from/to */
-  syncedRoundRange: [startRound: number, endRound: number]
-  /** The current detected tip of the configured Algorand blockchain. */
-  currentRound: number
-  /** The new watermark value to persist for the next call to
-   * `getSubscribedTransactions` to continue the sync.
-   * Will be equal to `syncedRoundRange[1]`. Only persist this
-   * after processing (or in the same atomic transaction as)
-   * subscribed transactions to keep it reliable. */
-  newWatermark: number
-  /** Any transactions that matched the given filter within
-   * the synced round range. This substantively uses the [indexer transaction
-   * format](https://developer.algorand.org/docs/rest-apis/indexer/#transaction)
-   * to represent the data with some additional fields.
-   */
-  subscribedTransactions: SubscribedTransaction[]
+/** Parameters to control a single subscription pull/poll. */
+export interface TransactionSubscriptionParams extends CoreTransactionSubscriptionParams {
+  /** The current round watermark that transactions have previously been synced to.
+   *
+   * Persist this value as you process transactions processed from this method
+   * to allow for resilient and incremental syncing.
+   *
+   * Syncing will start from `watermark + 1`.
+   *
+   * Start from 0 if you want to start from the beginning of time, noting that
+   * will be slow if `onMaxRounds` is `sync-oldest`.
+   **/
+  watermark: number
 }
 
-/** Configuration for a subscription */
-export interface SubscriptionConfig {
+/** Configuration for an `AlgorandSubscriber`. */
+export interface AlgorandSubscriberConfig extends CoreTransactionSubscriptionParams {
+  /** The set of filters to subscribe to / emit events for, along with optional data mappers. */
+  filters: SubscriberConfigFilter<unknown>[]
   /** The frequency to poll for new blocks in seconds; defaults to 1s */
   frequencyInSeconds?: number
   /** Whether to wait via algod `/status/wait-for-block-after` endpoint when at the tip of the chain; reduces latency of subscription */
   waitForBlockWhenAtTip?: boolean
-  /** The maximum number of rounds to sync at a time; defaults to 500 */
-  maxRoundsToSync?: number
-  /** Any ARC-28 event definitions to process from app call logs */
-  arc28Events?: Arc28EventGroup[]
-  /** The set of filters to subscribe to / emit events for */
-  filters: SubscriberConfigFilter<unknown>[]
-  /** The behaviour when the number of rounds to sync is greater than `maxRoundsToSync`:
-   *  * `skip-sync-newest`: Discard old rounds.
-   *  * `sync-oldest`: Sync from the oldest records up to `maxRoundsToSync` rounds.
-   *
-   *    **Note:** will be slow to catch up if sync is significantly behind the tip of the chain
-   *  * `sync-oldest-start-now`: Sync from the oldest records up to `maxRoundsToSync` rounds, unless
-   *    current watermark is `0` in which case it will start `maxRoundsToSync` back from the tip of the chain.
-   *  * `catchup-with-indexer`: Will catch up to `tipOfTheChain - maxRoundsToSync` using indexer (fast) and then
-   *    continue with algod.
-   *  * `fail`: Throw an error.
-   */
-  syncBehaviour: 'skip-sync-newest' | 'sync-oldest' | 'sync-oldest-start-now' | 'catchup-with-indexer' | 'fail'
   /** Methods to retrieve and persist the current watermark so syncing is resilient and maintains
    * its position in the chain */
   watermarkPersistence: {
