+++ tests/scenarios/subscriber.spec.ts
@@ -5,7 +5,7 @@ import { SendTransactionFrom } from '@algorandfoundation/algokit-utils/types/tra
 import { beforeEach, describe, test } from '@jest/globals'
 import { Algodv2, Indexer } from 'algosdk'
 import { AlgorandSubscriber } from '../../src'
-import { SubscriptionConfig } from '../../src/types/subscription'
+import { AlgorandSubscriberConfig } from '../../src/types'
 import { SendXTransactions } from '../transactions'
 import { waitFor } from '../wait'
 import { InMemoryWatermark } from '../watermarks'
@@ -23,7 +23,7 @@ describe('AlgorandSubscriber', () => {
   })
 
   const getSubscriber = (
-    config: { testAccount: SendTransactionFrom; configOverrides?: Partial<SubscriptionConfig>; initialWatermark?: number },
+    config: { testAccount: SendTransactionFrom; configOverrides?: Partial<AlgorandSubscriberConfig>; initialWatermark?: number },
     algod: Algodv2,
     indexer?: Indexer,
   ) => {
